requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
langgraph
langchain
langchain-openai
langchain-community
chromadb
duckdb
streamlit
pandas
numpy
requests
